vllm-ascend-chat/requirements.txt
vllm-ascend>=0.11.0rc1
gradio>=4.0.0
openai>=1.0.0
